Move metadata of source scenarios to includable module
